+++ program/actions/settings/about.php
@@ -36,6 +36,7 @@ class rcmail_action_settings_about extends rcmail_action
         $rcmail->output->add_handlers([
             'supportlink' => [$this, 'supportlink'],
             'pluginlist' => [$this, 'plugins_list'],
+            'skininfo' => [$this, 'skin_info'],
             'copyright' => static function () {
                 return 'Copyright &copy; 2005-2024, The Roundcube Dev Team';
             },
@@ -142,5 +143,20 @@ class rcmail_action_settings_about extends rcmail_action
         }
 
         return $table->show();
+    }
+
+    public static function skin_info($attrib)
+    {
+        $rcmail = rcmail::get_instance();
+        $meta = $rcmail->output->get_skin_info();
+
+        $content = html::p(null,
+            html::span('skinitem', html::span('skinname', rcube::Q($meta['name'])) . (!empty($meta['version']) ? '&nbsp;(' . $meta['version'] . ')' : '') . html::br() .
+                (!empty($meta['author_link']) ? html::span('skinauthor', $rcmail->gettext(['name' => 'skinauthor', 'vars' => ['author' => $meta['author_link']]])) . html::br() : '') .
+                (!empty($meta['license_link']) ? html::span('skinlicense', $rcmail->gettext('license') . ':&nbsp;' . $meta['license_link']) . html::br() : '') .
+                (!empty($meta['uri']) ? html::span('skinhomepage', $rcmail->gettext('source') . ':&nbsp;' . html::a(['href' => $meta['uri'], 'target' => '_blank', 'tabindex' => '-1'], rcube::Q($rcmail->gettext('download')))) : ''))
+        );
+
+        return $content;
     }
 }